### 2017-04-06 #225 @whedon
[PRE REVIEW]: getCRUCLdata: Use and Explore CRU CL v. 2.0 Climatology Elements in R
**Submitting author:** @adamhsparks (<a href="http://orcid.org/0000-0002-0061-8359">Adam H. Sparks</a>)
**Repository:** <a href="https://github.com/ropensci/getCRUCLdata" target ="_blank">https://github.com/ropensci/getCRUCLdata</a>
**Version:** 0.1.5
