### 2017-04-06 #225 @arfon
@noamross is this good to auto-accept into JOSS?
