### 2017-04-06 #225 @noamross
**getCRUCLdata** has been accepted to rOpenSci.  Reviews can be found here: https://github.com/ropensci/onboarding/issues/96
