### 2017-04-06 #230 @arfon
Accepting without additional review: https://github.com/ropensci/onboarding/issues/96
