### 2017-04-07 #230 @arfon
@adamhsparks your paper in now accepted into JOSS and you DOI is http://dx.doi.org/10.21105/joss.00230 ⚡️🚀💥
